[C8] Split self_intersect.go by phase (#2223)
self_intersect.go was 542 LOC, over the 500-line ceiling. Split it along
the three phases of the detector (the BVH was already in
self_intersect_bvh.go):

- self_intersect.go (147): broad phase + reporting — tessellate, box/BVH
  prune, shared-topology exclusion, witness reporting (SelfIntersection,
  SelfIntersections, sharedFaceBoundary, meshCrossesOffBoundary,
  faceMeshDeviation, onSharedBoundary).
- self_intersect_triangle.go (245): the narrow phase — the Möller straddle
  triangle-crossing test and its kind classification (triangleCrossing,
  signedDistances, intervalOverlap, unitCross, and the tol consts).
- self_intersect_coplanar.go (185): the coplanar branch — dominant-axis
  projection, separating-axis + Sutherland-Hodgman clip, area classification.

Behaviour-preserving: every function body is byte-identical, only moved;
imports narrowed per file. Self-intersection corpus + archguard pass before
and after with no golden updates. Net delta of recognizers / tolerance
constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2223.

(cherry picked from commit 69f100aba2d5e06d33ebffb4c0582f3051f2e745)
